authcore/pyproject.toml
[project]
name = "authcore"
version = "0.1.0"
description = "Standalone JWT authentication microservice"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "redis[hiredis]>=5.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    "bcrypt>=4.1.0",
    "dependency-injector>=4.41.0",
    "orjson>=3.9.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "aiosmtplib>=3.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "pyright>=1.1.0",
    "ruff>=0.4.0",
    "aiosqlite>=0.22.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=src/authcore --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src"]
exclude = ["alembic", ".venv", "tests"]
